Add Lists resource and CRUD actions/views
Create Lists resource with CRUD (create, read, update, delete) actions and views including index, show, new, edit, and partial form. Also, add tests for the Lists controller and model, as well as a system test for the Lists views using Capybara. These changes enable the creation and management of lists in the application.